Split dialogue parser into multiple files * Implemented ISerialisable for DialoguePlayback
